refactor: extract publishing flow service
